Change Keyword/Symbol modeling for better interop
This was a bit interesting - the instinct is to make Mal symbols just
be Ruby symbols, and that kinda works because Ruby symbols are kinda
used to represent methods, vars, etc in some contexts but actually
they are also used a lot like keywords - i.e. keys in Hash/Maps.

This becomes more important if you want to run some Mal code then use
the output of it (maybe a Hash/Map) in some (Dragon)Ruby code. An
obvious example is run some code then render the output as a sprite or
something - then your Keywords actually being strings means it doesn't
just work. That's annoying!

So here I've taken the approach that Keyword are Symbols, and
MalSymbols are objects wrapping Symbols, which means I can use the
MalSymbol type for making decisions. Anyway, it works.